requirements.txt
Django==1.7.1
Markdown==2.5.1
Werkzeug==0.9.6
dj-database-url==0.3.0
django-filter==0.8
djangorestframework==2.4.4
mock==1.0.1
nose==1.3.4
psycopg2==2.5.4
pyparsing==2.0.3
python-dateutil==2.2
requests==2.4.1
wsgiref==0.1.2
django-easy-pdf
xhtml2pdf>=0.0.6
reportlab>=2.7,<3